Boxing cleanup

+++ vpnservice/bgpmanager/bgpmanager-impl/src/main/java/org/opendaylight/netvirt/bgpmanager/ConfigureBgpCli.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2015 Ericsson India Global Services Pvt Ltd. and others.  All rights reserved.
+ * Copyright © 2015, 2017 Ericsson India Global Services Pvt Ltd. and others.  All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v1.0 which accompanies this distribution,
@@ -168,7 +168,7 @@ public class ConfigureBgpCli extends OsgiCommandSupport {
         }
         for (Neighbors nbr : nbrs) {
           if (nbrIp.equals(nbr.getAddress().getValue())) {
-             return nbr.getRemoteAs().longValue();
+             return nbr.getRemoteAs();
           }
         }
         return -1;
